requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
redis==5.0.1
supabase==2.4.0
websockets==12.0
python-socketio==5.11.1